A381

+++ B339.java
@@ -3,12 +3,12 @@ import java.util.*;
 public class B339 {
 	public static void main(String[] args) {
 		Scanner sc = new Scanner(System.in);
-		int n = sc.nextInt(); // number of houses
+		long n = sc.nextLong(); // number of houses
 		int m = sc.nextInt(); // m things to do
 		
-		int arr[] = new int[m]; // creating array
+		long arr[] = new long[m]; // creating array
 			
-		int dist = 1; // initially at house 1
+		long dist = 1; // initially at house 1
 		for ( int i = 0 ; i < m ; i++) {
 			int x = sc.nextInt();
 			arr[i] = x;
